+++ aim/ql/tree/abstract_syntax_tree.py
@@ -6,7 +6,7 @@ from aim.ql.tokens.token import Token
 
 class AbstractSyntaxTree(Tree):
     def build_from_expression(self, expression):
-        token = expression.node
+        token = expression.node if hasattr(expression, 'node') else expression
         self.head = self.NODE(token, 0, 0)
         if isinstance(token.value, list):
             self.append_tokens(token.value, self.head)